3DS models moved to separate package
