+++ checkout-3ds/src/main/kotlin/com/processout/sdk/checkout/threeds/POCheckout3DSHandler.kt
@@ -11,10 +11,10 @@ import com.checkout.threeds.standalone.models.AuthenticationRequestParameters
 import com.checkout.threeds.standalone.models.ConfigParameters
 import com.checkout.threeds.standalone.models.DirectoryServerData
 import com.checkout.threeds.standalone.models.StandaloneResult
-import com.processout.sdk.api.model.request.PO3DS2AuthenticationRequest
-import com.processout.sdk.api.model.response.PO3DS2Challenge
-import com.processout.sdk.api.model.response.PO3DS2Configuration
-import com.processout.sdk.api.model.response.PO3DSRedirect
+import com.processout.sdk.api.model.threeds.PO3DS2AuthenticationRequest
+import com.processout.sdk.api.model.threeds.PO3DS2Challenge
+import com.processout.sdk.api.model.threeds.PO3DS2Configuration
+import com.processout.sdk.api.model.threeds.PO3DSRedirect
 import com.processout.sdk.api.service.PO3DSHandler
 import com.processout.sdk.api.service.PO3DSResult
 import com.processout.sdk.checkout.threeds.POCheckout3DSHandlerState.*
