+++ sdk/src/main/kotlin/com/processout/sdk/api/service/ThreeDSServiceImpl.kt
@@ -1,12 +1,12 @@
 package com.processout.sdk.api.service
 
 import android.util.Base64
-import com.processout.sdk.api.model.request.PO3DS2AuthenticationRequest
-import com.processout.sdk.api.model.response.PO3DS2Challenge
-import com.processout.sdk.api.model.response.PO3DS2Configuration
-import com.processout.sdk.api.model.response.PO3DSRedirect
 import com.processout.sdk.api.model.response.POCustomerAction
 import com.processout.sdk.api.model.response.POCustomerAction.Type.*
+import com.processout.sdk.api.model.threeds.PO3DS2AuthenticationRequest
+import com.processout.sdk.api.model.threeds.PO3DS2Challenge
+import com.processout.sdk.api.model.threeds.PO3DS2Configuration
+import com.processout.sdk.api.model.threeds.PO3DSRedirect
 import com.processout.sdk.core.POFailure
 import com.squareup.moshi.JsonClass
 import com.squareup.moshi.Moshi
